change: move sync sub-command to under the kvs command #46
